divide models by folders
