+++ Src/Campus.Master.API/UI/src/app/data/dao/interface/CategoryDAO.ts
@@ -1,6 +1,6 @@
 import { CommonDAO } from './CommonDAO';
 import { Observable } from 'rxjs';
-import { Project } from '../../../shared/models/project';
+import { Project } from '../../../shared/models/task-list/project';
 
 export interface CategoryDAO extends CommonDAO<Project> {
