+++ Src/Campus.Master.API/UI/src/app/data/dao/interface/PriorityDAO.ts
@@ -1,5 +1,5 @@
 import { CommonDAO } from './CommonDAO';
-import { Priority } from '../../../shared/models/priority';
+import { Priority } from '../../../shared/models/task-list/priority';
 
 export interface PriorityDAO extends CommonDAO<Priority> {
 }
