+++ Src/Campus.Master.API/UI/src/app/data/dao/interface/StatusDAO.ts
@@ -1,5 +1,5 @@
 import { CommonDAO } from './CommonDAO';
-import { Status } from '../../../shared/models/status';
+import { Status } from '../../../shared/models/task-list/status';
 
 export interface StatusDAO extends CommonDAO<Status> {
