+++ Src/Campus.Master.API/UI/src/app/data/testData.ts
@@ -1,7 +1,7 @@
-import { Priority } from '../shared/models/priority';
-import { Task } from '../shared/models/task';
-import { Project } from '../shared/models/project';
-import { Status } from '../shared/models/status';
+import { Priority } from '../shared/models/task-list/priority';
+import { Task } from '../shared/models/task-list/task';
+import { Project } from '../shared/models/task-list/project';
+import { Status } from '../shared/models/task-list/status';
 
 export class TestData {
