+++ Src/Campus.Master.API/UI/src/app/shared/models/state-transfer.ts
@@ -1,0 +1,4 @@
+export interface StateTransfer {
+    message: string;
+    payload: string;
+}
