+++ Src/Campus.Master.API/UI/src/app/shared/models/user/register-user.ts
@@ -1,0 +1,8 @@
+export interface RegisterUser {
+    password: string;
+    confirmPassword: string;
+    email: string;
+    firstName: string;
+    lastName: string;
+    gender: number;
+}
