+++ Src/Campus.Master.API/UI/src/app/shared/models/user/user.ts
@@ -1,0 +1,4 @@
+export interface User {
+    email: string;
+    password: string;
+  }
