+++ Src/Campus.Master.API/UI/src/app/shared/services/project.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Project } from 'src/app/shared/models/project';
+import { Project } from 'src/app/shared/models/task-list/project';
 import { HttpClient } from '@angular/common/http';
 import { environment } from 'src/environments/environment';
-import { StateTransfer } from '../interfaces';
+import { StateTransfer } from '../models/state-transfer';
 
 @Injectable({
   providedIn: 'root'
