+++ Src/Campus.Master.API/UI/src/app/shared/services/registration.service.ts
@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { RegisterUser, StateTransfer } from '../interfaces';
 import { environment } from 'src/environments/environment';
+import { RegisterUser } from '../models/user/register-user';
+import { StateTransfer } from '../models/state-transfer';
 
 @Injectable({
   providedIn: 'root',
