+++ Src/Campus.Master.API/UI/src/app/views/tasks_board/task-board-layout/task-board-layout.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { Task } from '../../../shared/models/task';
+import { Task } from '../../../shared/models/task-list/task';
 import { DataHandlerService } from '../../../shared/services/data-handler.service';
-import { Project } from '../../../shared/models/project';
+import { Project } from '../../../shared/models/task-list/project';
 import { zip } from 'rxjs';
 
 @Component({
