+++ Src/Campus.Master.API/UI/src/app/views/tasks_board/task/task.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnChanges, OnInit, Output } from '@angular/core';
-import { Task } from 'src/app/shared/models/task';
+import { Task } from 'src/app/shared/models/task-list/task';
 import { DataHandlerService } from '../../../shared/services/data-handler.service';
 import { SortDateService } from '../../../shared/services/sort-date.service';
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
